service_b/pyproject.toml
[project]
name = "service-b"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ninja>=1.3.0",
    "requests>=2.32.3",
]
